adjust tests, increase coverage
