+++ setup.cfg
@@ -57,7 +57,6 @@ exclude =
     mitmproxy/master.py
     mitmproxy/net/check.py
     mitmproxy/net/http/cookies.py
-    mitmproxy/net/http/headers.py
     mitmproxy/net/http/message.py
     mitmproxy/net/http/multipart.py
     mitmproxy/net/tcp.py
