+++ test/mitmproxy/proxy/test_context.py
@@ -1,82 +1,5 @@
-from mitmproxy import connection
 from mitmproxy.proxy import context
 from mitmproxy.test import tflow, taddons
-
-
-class TestConnection:
-    def test_basic(self):
-        c = connection.Client(
-            ("127.0.0.1", 52314),
-            ("127.0.0.1", 8080),
-            1607780791
-        )
-        assert not c.tls_established
-        c.timestamp_tls_setup = 1607780792
-        assert c.tls_established
-        assert c.connected
-        c.state = connection.ConnectionState.CAN_WRITE
-        assert not c.connected
-
-    def test_eq(self):
-        c = tflow.tclient_conn()
-        c2 = c.copy()
-        assert c == c
-        assert c != c2
-        assert c != 42
-        assert hash(c) != hash(c2)
-
-        c2.id = c.id
-        assert c == c2
-
-
-class TestClient:
-    def test_basic(self):
-        c = connection.Client(
-            ("127.0.0.1", 52314),
-            ("127.0.0.1", 8080),
-            1607780791
-        )
-        assert repr(c)
-        assert str(c)
-        c.timestamp_tls_setup = 1607780791
-        assert str(c)
-        c.alpn = b"foo"
-        assert str(c) == "Client(127.0.0.1:52314, state=open, alpn=foo)"
-
-    def test_state(self):
-        c = tflow.tclient_conn()
-        assert connection.Client.from_state(c.get_state()).get_state() == c.get_state()
-
-        c2 = tflow.tclient_conn()
-        assert c != c2
-
-        c2.timestamp_start = 42
-        c.set_state(c2.get_state())
-        assert c.timestamp_start == 42
-
-        c3 = c.copy()
-        assert c3.get_state() != c.get_state()
-        c.id = c3.id = "foo"
-        assert c3.get_state() == c.get_state()
-
-
-class TestServer:
-    def test_basic(self):
-        s = connection.Server(("address", 22))
-        assert repr(s)
-        assert str(s)
-        s.timestamp_tls_setup = 1607780791
-        assert str(s)
-        s.alpn = b"foo"
-        s.sockname = ("127.0.0.1", 54321)
-        assert str(s) == "Server(address:22, state=closed, alpn=foo, src_port=54321)"
-
-    def test_state(self):
-        c = tflow.tserver_conn()
-        c2 = c.copy()
-        assert c2.get_state() != c.get_state()
-        c.id = c2.id = "foo"
-        assert c2.get_state() == c.get_state()
 
 
 def test_context():
